+++ gs-blog.php
@@ -49,4 +49,6 @@ queue_style ('font_awesome', GSBACK );
 
 # Register/Queue Scripts
 register_script('pluginManagementFA', $SITEURL.'plugins/'.BLOGFILE.'/js/pluginManagementFA.js', '1.0', TRUE);
+register_script('table_paging', $SITEURL.'plugins/'.BLOGFILE.'/js/paging.js', '1.0', FALSE);
 queue_script('pluginManagementFA', GSBACK);
+queue_script('table_paging', GSBACK);